Two Point Equal Interval Search Method
This code is written in Python 2.7.3 on Ubuntu 12.04 LTS using spyder IDE. The code depicts a one-dimensional optimization technique named as "Two Point Equal Interval Search Method" for locating a local minima in a continuous and unimodal function within a specified closed interval. Any desired function can be specified within the user-defined function 'equation()' in the code. This code requires upper and lower limits of the closed interval along with the final interval of uncertainty in percentage(error) as inputs. Minima, value of the function at minima, number of iterations required and a plot of the specified function within the specified closed interval(for verification) are given as the outputs.