Golden Section Search
This code is written in Python 2.7.3 using spyder IDE on Ubuntu 12.04 LTS. This code employs the Golden Section Search for locating a local minima in a function specified by the user in the user-defined 'equation()' function in the required closed interval. End points of the closed interval and final interval of uncertainty are required as inputs. Output is generated in the form of number of iterations required and the x and f(x) values corresponding to the local minima. This code also generates a table showing the interval end points after every iteration along with a plot of the function for verification purposes.